pyproject.toml
[project]
name="eth-ue"
version="0.0.1"
dependencies = [
    "comet-ml>=3.49.1",
    "hydra-core>=1.3.2",
    "joblib>=1.4.2",
    "lightning>=2.5.0.post0",
    "mypy>=1.15.0",
    "numpy>=2.2.2",
    "pandas>=2.2.3",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.6",
    "torch>=2.6.0",
]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E701", "E203"]
max-doc-length = 120
exclude = "test_*.py"

[tool.pytest.ini_options]
timeout = 60
minversion = "6.0"
python_files = "test_*.py"
addopts = "-p no:requests_mock"
testpaths = ["tests", ".tmp"]


[tool.mypy]
no_incremental = true
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_any_generics = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
allow_untyped_decorators = true
enable_incomplete_feature = "NewGenericSyntax"
exclude = ["tests/", ".tmp/", ".venv/"]

[tool.pyright]
include = ["checker"]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    ".venv",
    "venv",
    ".tmp",
    "tests",
]


[tool.ruff]
target-version = "py311"

# Enable the pycodestyle (`E`) and Pyflakes (`F`) rules by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
lint.select = ["E", "F"]
lint.ignore = ["E203"]

# Allow autofix for all enabled rules (when `--fix`) is provided.
lint.fixable = ["ALL"]
lint.unfixable = []

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",

    "tests",
    ".tmp",
]

line-length = 88
